# docs/swagger/bloodconnect.yaml
openapi: 3.0.3
info:
  title: BloodConnect API
  description: |
    API-first contract for BloodConnect REST API.

    This document is the shared source of truth for Backend, Flutter, Postman,
    Swagger UI, and future client SDK generation.
  version: 1.0.0
  contact:
    name: BloodConnect API Team
  license:
    name: Proprietary
    url: https://bloodconnect.id/license
servers:
  - url: /api/v1
    description: Base API path
security:
  - bearerAuth: []
...
paths:
...
  /screenings:
    post:
      tags:
        - Screening
      summary: Submit donor screening
      description: Submits eligibility screening answers for a donor before responding to a blood request.
      operationId: createScreening
      requestBody:
        $ref: "#/components/requestBodies/CreateScreeningBody"
      responses:
        "201":
          description: Screening submitted successfully.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessResponse"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/ScreeningResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "422":
          $ref: "#/components/responses/ValidationError"
        "500":
          $ref: "#/components/responses/InternalServerError"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT access token using the Authorization header.
        Example: `Authorization: Bearer eyJhbGciOiJIUzI1NiIs...`
...
  requestBodies:
...
    CreateScreeningBody:
      required: true
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/CreateScreeningRequest"
...
  responses:
...
    BadRequest:
      description: Bad request.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          examples:
            badRequest:
              $ref: "#/components/examples/BadRequestExample"
    Unauthorized:
      description: Unauthorized request.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          examples:
            unauthorized:
              $ref: "#/components/examples/UnauthorizedExample"
...
    ValidationError:
      description: Validation failed.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ValidationErrorResponse"
          examples:
            validationError:
              $ref: "#/components/examples/ValidationErrorExample"
    InternalServerError:
      description: Internal server error.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          examples:
            internalServerError:
              $ref: "#/components/examples/InternalServerErrorExample"
  schemas:
    Gender:
      type: string
      enum:
        - Male
        - Female
    BloodType:
      type: string
      enum:
        - A
        - B
        - AB
        - O
    Rhesus:
      type: string
      enum:
        - "+"
        - "-"
...
    CreateScreeningRequest:
      type: object
      required:
        - blood_request_id
        - weight_ok
        - healthy
        - taking_medicine
        - pregnant
        - tattoo
        - operation_history
      properties:
        blood_request_id:
          type: string
          format: uuid
        last_donor_date:
          type: string
          format: date
          nullable: true
        weight_ok:
          type: boolean
        healthy:
          type: boolean
        taking_medicine:
          type: boolean
        pregnant:
          type: boolean
        tattoo:
          type: boolean
        operation_history:
          type: boolean
...
    UserResponse:
      type: object
      required:
        - id
        - full_name
        - email
        - phone
        - blood_type
        - rhesus
        - is_available
        - is_verified
        - is_active
        - created_at
        - updated_at
      properties:
        id:
          type: string
          format: uuid
        full_name:
          type: string
          example: Andi Pratama
        email:
          type: string
          format: email
        phone:
          type: string
        gender:
          $ref: "#/components/schemas/Gender"
        birth_date:
          type: string
          format: date
          nullable: true
        blood_type:
          $ref: "#/components/schemas/BloodType"
        rhesus:
          $ref: "#/components/schemas/Rhesus"
        weight:
          type: integer
          nullable: true
        last_donor_date:
          type: string
          format: date
          nullable: true
        latitude:
          type: number
          format: double
          nullable: true
        longitude:
          type: number
          format: double
          nullable: true
        profile_photo:
          type: string
          format: uri
          nullable: true
        is_available:
          type: boolean
        is_verified:
          type: boolean
        is_active:
          type: boolean
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    ScreeningResponse:
      type: object
      required:
        - id
        - user_id
        - blood_request_id
        - weight_ok
        - healthy
        - taking_medicine
        - pregnant
        - tattoo
        - operation_history
        - eligible
        - created_at
        - updated_at
      properties:
        id:
          type: string
          format: uuid
        user_id:
          type: string
          format: uuid
        blood_request_id:
          type: string
          format: uuid
        last_donor_date:
          type: string
          format: date
          nullable: true
        weight_ok:
          type: boolean
        healthy:
          type: boolean
        taking_medicine:
          type: boolean
        pregnant:
          type: boolean
        tattoo:
          type: boolean
        operation_history:
          type: boolean
        eligible:
          type: boolean
        user:
          $ref: "#/components/schemas/UserResponse"
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    SuccessResponse:
      type: object
      required:
        - success
        - message
        - data
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: Operation completed successfully
        data:
          type: object
          nullable: true
    ErrorResponse:
      type: object
      required:
        - success
        - message
        - errors
      properties:
        success:
          type: boolean
          example: false
        message:
          type: string
          example: Request failed
        errors:
          type: object
          nullable: true
          additionalProperties: true
    ValidationErrorResponse:
      type: object
      required:
        - success
        - message
        - errors
      properties:
        success:
          type: boolean
          example: false
        message:
          type: string
          example: Validation failed
        errors:
          type: object
          additionalProperties:
            type: array
            items:
              type: string
  examples:
...
    BadRequestExample:
      summary: 400 Bad Request
      value:
        success: false
        message: Invalid request body
        errors:
          body:
            - Request body must be valid JSON
    UnauthorizedExample:
      summary: 401 Unauthorized
      value:
        success: false
        message: Unauthorized
        errors:
          token:
            - Token is missing, invalid, or expired
...
    ValidationErrorExample:
      summary: 422 Unprocessable Entity
      value:
        success: false
        message: Validation failed
        errors:
          email:
            - Email is required
          password:
            - Password must be at least 8 characters
    InternalServerErrorExample:
      summary: 500 Internal Server Error
      value:
        success: false
        message: Internal server error
        errors:
          server:
            - An unexpected error occurred